Moved around the action types
