+++ client/src/components/scene/zoneScenesComponent.tsx
@@ -20,8 +20,7 @@ along with Home Lights.  If not, see <http://www.gnu.org/licenses/>.
 import { Divider, List } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import React, { Fragment, FunctionComponent } from 'react';
-import { Light, Scene, Zone, ZoneState } from '../../common/types';
-import { EditMode } from '../../types';
+import { EditMode, Light, Scene, Zone, ZoneState } from '../../common/types';
 import { CreateSceneButtonContainer } from './createSceneButtonContainer';
 import { SceneComponent, SceneComponentDispatch } from './sceneComponent';
