+++ client/src/components/zone/zoneComponent.tsx
@@ -28,8 +28,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import { ExpandMore as ExpandMoreIcon } from '@material-ui/icons';
 import React, { FunctionComponent } from 'react';
 import { BRIGHTNESS_STEP, MAX_BRIGHTNESS } from '../../common/config';
-import { Scene, Zone, ZoneState } from '../../common/types';
-import { EditMode } from '../../types';
+import { EditMode, Scene, Zone, ZoneState } from '../../common/types';
 import { useContentStyles } from '../lib/pageStyles';
 import { ZoneScenesContainer } from '../scene/zoneScenesContainer';
 import { DeleteZoneButton, DeleteZoneButtonDispatch } from './deleteZoneButton';
